scripts/requirements.txt
PyGithub
GitPython
boltons
jinja2
pyyaml
pillow
